+++ lib/typed_struct.ex
@@ -2,6 +2,7 @@
 # SPDX-FileCopyrightText: 2018 Marcin Górnik <marcin.gornik@gmail.com>
 # SPDX-FileCopyrightText: 2022 Jonathan Chukinas <chukinas@gmail.com>
 # SPDX-FileCopyrightText: 2022 Balázs Jávorszky <javorszky.balazs@estyle.hu>
+# SPDX-FileCopyrightText: 2022 Phil Chen <06fahchen@gmail.com>
 #
 # SPDX-License-Identifier: MIT
 
@@ -15,9 +16,11 @@ defmodule TypedStruct do
   @accumulating_attrs [
     :ts_plugins,
     :ts_plugin_fields,
+    :ts_parameters,
+    :ts_parameter_docs,
     :ts_fields,
     :ts_types,
-    :ts_docs,
+    :ts_field_docs,
     :ts_enforce_keys
   ]
 
@@ -70,6 +73,18 @@ defmodule TypedStruct do
         end
       end
 
+  You can also add type parameters:
+
+      defmodule MyModule do
+        use TypedStruct
+
+        typedstruct do
+          parameter :type
+
+          field :field, type
+        end
+      end
+
   You can create the struct in a submodule instead:
 
       defmodule MyModule do
@@ -118,36 +133,48 @@ defmodule TypedStruct do
       @enforce_keys @ts_enforce_keys
       defstruct @ts_fields
 
-      TypedStruct.__typedoc__(@ts_docs)
-      TypedStruct.__type__(@ts_types, unquote(opts))
+      TypedStruct.__typedoc__(@ts_parameter_docs, @ts_field_docs)
+
+      TypedStruct.__type__(
+        Enum.reverse(@ts_parameters),
+        @ts_types,
+        unquote(opts)
+      )
     end
   end
 
   @doc false
-  defmacro __typedoc__(docs) do
-    quote bind_quoted: [docs: docs] do
-      field_docs =
-        docs
+  defmacro __typedoc__(parameter_docs, field_docs) do
+    quote bind_quoted: [
+            parameter_docs: parameter_docs,
+            field_docs: field_docs
+          ] do
+      parameter_docs =
+        parameter_docs
         |> Enum.reverse()
         |> Enum.filter(fn {_, doc} -> !is_nil(doc) end)
-        |> Enum.map(fn {name, doc} -> "- `#{name}` - #{doc}" end)
+        |> Enum.map_join("\n", fn {name, doc} -> "- `#{name}` - #{doc}" end)
+        |> TypedStruct.__add_heading__("Type parameters")
 
-      if field_docs != [] do
+      field_docs =
+        field_docs
+        |> Enum.reverse()
+        |> Enum.filter(fn {_, doc} -> !is_nil(doc) end)
+        |> Enum.map_join("\n", fn {name, doc} -> "- `#{name}` - #{doc}" end)
+        |> TypedStruct.__add_heading__("Fields")
+
+      if parameter_docs != "" || field_docs != "" do
         # If there are field docs, we complete the `@typedoc` with field
         # documentation. However, if there is no `@typedoc` already, let’s emit
         # a warning instead.
         if Module.has_attribute?(__MODULE__, :typedoc) do
           @typedoc """
-          #{@typedoc}
-
-          ## Fields
-
-          #{Enum.join(field_docs, "\n")}
+          #{@typedoc}#{parameter_docs}#{field_docs}
           """
         else
           IO.warn(
             """
-            adding field documentation has no effect without a @typedoc
+            adding parameter or field documentation has no effect without a @typedoc
 
             hint: add a @typedoc on your `typedstruct` definition
             """,
@@ -159,14 +186,22 @@ defmodule TypedStruct do
   end
 
   @doc false
-  defmacro __type__(types, opts) do
+  def __add_heading__("", _heading), do: ""
+  def __add_heading__(doc, heading), do: "\n\n## #{heading}\n\n#{doc}"
+
+  @doc false
+  defmacro __type__(parameters, types, opts) do
     if Keyword.get(opts, :opaque, false) do
-      quote bind_quoted: [types: types] do
-        @opaque t() :: %__MODULE__{unquote_splicing(types)}
+      quote bind_quoted: [parameters: parameters, types: types] do
+        @opaque t(unquote_splicing(parameters)) :: %__MODULE__{
+                  unquote_splicing(types)
+                }
       end
     else
-      quote bind_quoted: [types: types] do
-        @type t() :: %__MODULE__{unquote_splicing(types)}
+      quote bind_quoted: [parameters: parameters, types: types] do
+        @type t(unquote_splicing(parameters)) :: %__MODULE__{
+                unquote_splicing(types)
+              }
       end
     end
   end
@@ -197,6 +232,36 @@ defmodule TypedStruct do
       require unquote(plugin)
       unquote(plugin).init(unquote(opts))
     end
+  end
+
+  @doc """
+  Defines a type parameter for the currently defined struct.
+
+  ## Example
+
+      typedstruct do
+        # Defines a type parameter named `type_param`
+        parameter :type_param
+
+        # The type parameter can be used as a type in the `field` macro.
+        field :a_field, type_param
+      end
+  """
+  defmacro parameter(name, opts \\ []) do
+    quote bind_quoted: [name: name, opts: opts] do
+      TypedStruct.__parameter__(name, opts, __ENV__)
+    end
+  end
+
+  @doc false
+  def __parameter__(name, opts, %Macro.Env{module: mod}) when is_atom(name) do
+    Module.put_attribute(mod, :ts_parameters, Macro.var(name, mod))
+    Module.put_attribute(mod, :ts_parameter_docs, {name, opts[:doc]})
+  end
+
+  def __parameter__(name, _opts, _env) do
+    raise ArgumentError,
+          "the name of a type parameter must be an atom, got #{inspect(name)}"
   end
 
   @doc """
@@ -247,14 +312,14 @@ defmodule TypedStruct do
     Module.put_attribute(mod, :ts_fields, {name, opts[:default]})
     Module.put_attribute(mod, :ts_plugin_fields, {name, type, opts, env})
     Module.put_attribute(mod, :ts_types, {name, type_for(type, nullable?)})
-    Module.put_attribute(mod, :ts_docs, {name, opts[:doc]})
+    Module.put_attribute(mod, :ts_field_docs, {name, opts[:doc]})
     if enforce?, do: Module.put_attribute(mod, :ts_enforce_keys, name)
   end
 
   # Checks whether some value looks like Elixir AST.
   defp ast?({name, meta, params})
        when (is_atom(name) or is_tuple(name)) and is_list(meta) and
-              is_list(params),
+              (is_list(params) or is_nil(params)),
        do: true
 
   defp ast?(_), do: false